requirements.txt
Pillow
music21==9.1.0

